XW-1104 | renamed /models/mediawiki to /models/wiki
