+++ front/main/app/models/wiki/article.js
