+++ front/main/app/routes/application.js
@@ -1,6 +1,6 @@
 import Ember from 'ember';
 import TrackClickMixin from '../mixins/track-click';
-import ArticleModel from '../models/mediawiki/article';
+import ArticleModel from '../models/wiki/article';
 import {activate as variantTestingActivate} from 'common/utils/variantTesting';
 import {getLinkInfo} from 'common/utils/articleLink';
 import {normalizeToUnderscore} from 'common/utils/string';
